Fixes Kirby being unable to cancel Rollout
