+++ src/fighter/kirby/status.rs
@@ -1,4 +1,5 @@
 mod kirby;
+mod purin;
 mod ganon;
 mod lucario;
 mod ike;
@@ -6,6 +7,7 @@ mod belmont;
 
 pub fn install() {
     kirby::install();
+    purin::install();
     ganon::install();
     lucario::install();
     ike::install();
